Decodable test added and edited
